Close the fibCounter when not needed, fix #2
The fibCounter has been amended to receive the amount of values that
need to be calculated. After the required amount has been calculated
and sent back to the channel, the channel is closed and the routine ends.
For low n values this speeds up the calculation multiple times (e.g. for
`n=10`, the benchmark runs 5x faster).